+++ docs/user/changes.rst
@@ -5,7 +5,7 @@ Release history
 - feat: new :meth:`api.telephony.get_large_organization_status <wxc_sdk.telephony.TelephonyApi.get_large_organization_status>`
 - feat: new cdr fields interaction_id, wx_cc_consult_merge_status
 - feat: new :meth:`api.person_settings.app_shared_line.members_count <wxc_sdk.person_settings.app_shared_line.AppSharedLineApi.members_count>`
-
+- feat: new api :attr:`api.telephony.call_controls_members <wxc_sdk.telephony.TelephonyApi.call_controls_members>`
 
 1.29.0
 ------